remove duplicate code from banners
